Fix FruSize D-Bus type to support values above 255
Change FruSize parsing in DoFruConfig from uint8_t
to uint16_t so values <= 512 are represented
correctly on D-Bus.

+++ src/fru_device/fru_device.cpp
@@ -1374,7 +1374,7 @@ void DoFruConfig(sdbusplus::asio::object_server& objServer)
         uint8_t addrConf = static_cast<uint8_t>(fruConfigs["Address"]);
         std::string fruName = static_cast<std::string>(fruConfigs["Name"]);
         uint8_t fruId = static_cast<uint8_t>(fruConfigs["FruId"]);
-        uint8_t fruSize = static_cast<uint8_t>(fruConfigs["FruSize"]);
+        uint16_t fruSize = static_cast<uint16_t>(fruConfigs["FruSize"]);
         std::shared_ptr<sdbusplus::asio::dbus_interface> iface =
             objServer.add_interface(
                 "/xyz/openbmc_project/FruDevice/" + fruName,
